Update README.md
Session 2 – Introduction to R

ASSIGNMENT -1


Problem Statement

1.	What are the different methods to call a function in R?

Answer: 
3 ways – call () , do.call () & recall ()  


2.	The lazy evaluation of a function means, the argument is evaluated only if it is evaluated only if it is used the body of the function. Say True or False 

Answer:  True


3. State True or False:

a. Insights driven from descriptive analytics is not meaningful. - False

b. The number of values in each Elements of a list, should be equal. - True

c. The datasets are not stored in memory of the computer using R. - True

d. Data frames and matrices are two dimensional however the array is multidimensional. - True